refactor: move header from pkg to libs
