+++ header/verify.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/tendermint/tendermint/light"
 
-	headerpkg "github.com/celestiaorg/celestia-node/pkg/header"
+	headerpkg "github.com/celestiaorg/celestia-node/libs/header"
 )
 
 // TODO(@Wondertan): We should request TrustingPeriod from the network's state params or
