+++ libs/header/hash.go
