+++ libs/header/p2p/subscription.go
@@ -7,7 +7,7 @@ import (
 
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 
-	"github.com/celestiaorg/celestia-node/pkg/header"
+	"github.com/celestiaorg/celestia-node/libs/header"
 )
 
 // subscription handles retrieving ExtendedHeaders from the header pubsub topic.
