+++ libs/header/store/store_test.go
@@ -12,7 +12,7 @@ import (
 
 	tmrand "github.com/tendermint/tendermint/libs/rand"
 
-	"github.com/celestiaorg/celestia-node/pkg/header"
+	"github.com/celestiaorg/celestia-node/libs/header"
 )
 
 func TestStore(t *testing.T) {
