+++ libs/header/sync/sync_head.go
@@ -6,7 +6,7 @@ import (
 
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 
-	"github.com/celestiaorg/celestia-node/pkg/header"
+	"github.com/celestiaorg/celestia-node/libs/header"
 )
 
 // Head returns the Syncer's latest known header. It calls 'networkHead' in order to
