+++ libs/header/sync/sync_test.go
@@ -9,9 +9,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/celestiaorg/celestia-node/pkg/header"
-	"github.com/celestiaorg/celestia-node/pkg/header/local"
-	"github.com/celestiaorg/celestia-node/pkg/header/store"
+	"github.com/celestiaorg/celestia-node/libs/header"
+	"github.com/celestiaorg/celestia-node/libs/header/local"
+	"github.com/celestiaorg/celestia-node/libs/header/store"
 )
 
 func TestSyncSimpleRequestingHead(t *testing.T) {
