+++ nodebuilder/tests/swamp/swamp.go
@@ -21,6 +21,7 @@ import (
 	"github.com/celestiaorg/celestia-node/core"
 	"github.com/celestiaorg/celestia-node/header"
 	headercore "github.com/celestiaorg/celestia-node/header/core"
+	headerpkg "github.com/celestiaorg/celestia-node/libs/header"
 	"github.com/celestiaorg/celestia-node/libs/keystore"
 	"github.com/celestiaorg/celestia-node/logs"
 	"github.com/celestiaorg/celestia-node/nodebuilder"
@@ -28,7 +29,6 @@ import (
 	"github.com/celestiaorg/celestia-node/nodebuilder/node"
 	"github.com/celestiaorg/celestia-node/nodebuilder/p2p"
 	"github.com/celestiaorg/celestia-node/nodebuilder/state"
-	headerpkg "github.com/celestiaorg/celestia-node/pkg/header"
 )
 
 var blackholeIP6 = net.ParseIP("100::")
